move Api.Request and Api.Response to their own file
