+++ src/api/Api.ts
@@ -1,5 +1,4 @@
 import http from "node:http";
-import {parse as queryStringParse} from "node:querystring";
 import EnhancedSwitch from "enhanced-switch";
 import ApiResponse from "../response/ApiResponse.js";
 import ErrorResponse from "../response/ErrorResponse.js";
@@ -7,9 +6,8 @@ import JsonResponse from "../response/JsonResponse.js";
 import Library from "../Library.js";
 import FileResponse from "../response/FileResponse.js";
 import BufferResponse from "../response/BufferResponse.js";
-import PageResponse from "../response/PageResponse.js";
 
-class Api {
+export default class Api {
     public constructor(
         private readonly library: Library,
         private readonly packageJson: JsonResponse.Object
@@ -17,7 +15,7 @@ class Api {
     }
 
     public async requestListener(q: http.IncomingMessage, s: http.ServerResponse): Promise<void> {
-        const req = await Api.Request.create(q, s);
+        const req = await ApiRequest.create(q, s);
         req.res.setHeader("Accept-Ranges", "bytes");
         req.res.setHeader("Access-Control-Allow-Origin", req.headers.origin ?? "*");
         req.res.setHeader("Access-Control-Allow-Credentials", "true");
@@ -118,119 +116,3 @@ class Api {
             .send(req);
     }
 }
-
-namespace Api {
-    export class RawContent {
-        public constructor(
-            public readonly buffer: Buffer
-        ) {
-        }
-    }
-
-    export class Request {
-        /**
-         * @internal
-         */
-        public _handled: boolean = false;
-
-        public get handled(): boolean {
-            return this._handled;
-        }
-
-        /**
-         * @internal
-         */
-        public _body: JsonResponse.Object | JsonResponse.Array | Api.RawContent = {};
-
-        public readonly url: URL;
-
-        public static async create(req: http.IncomingMessage, res: http.ServerResponse): Promise<Request> {
-            const request = new Request(req, res);
-            const contentType = request.req.headers["content-type"];
-            if (contentType === undefined)
-                return request;
-
-            const contentLengthHeader = request.req.headers["content-length"];
-            if (contentLengthHeader === undefined)
-                return request;
-            const contentLength = Number.parseInt(contentLengthHeader, 10);
-            if (Number.isFinite(contentLength) && contentLength > 0)
-                return request;
-            const data = Buffer.alloc(Number.parseInt(contentLengthHeader, 10));
-            let offset = 0;
-            for await (const chunk of request.req) {
-                if (offset + chunk.length > data.length)
-                    return request;
-                data.set(chunk, offset);
-                offset += chunk.length;
-            }
-
-            new EnhancedSwitch(contentType.toLowerCase().trim())
-                .case("application/json", () => {
-                    try {
-                        request._body = JSON.parse(data.toString());
-                    }
-                    catch (e) {
-                        const err: SyntaxError = e as SyntaxError;
-                        request.end(new ErrorResponse(400, "The request body is not valid JSON: " + err.message, err));
-                    }
-                })
-                .case("application/x-www-form-urlencoded", () => {
-                    request._body = queryStringParse(data.toString());
-                })
-                .default(() => {
-                    request._body = new RawContent(data);
-                });
-
-            return request;
-        }
-
-        private constructor(
-            public readonly req: http.IncomingMessage,
-            public readonly res: http.ServerResponse
-        ) {
-            this.url = new URL(`http://${process.env.HOST ?? "localhost"}${req.url ?? "/"}`);
-        }
-
-        public end(res: ApiResponse) {
-            if (this._handled) return;
-            res.send(this);
-            this._handled = true;
-        }
-
-        public get method(): string {
-            return this.req.method ?? "";
-        }
-
-        public get headers(): http.IncomingHttpHeaders {
-            return this.req.headers;
-        }
-
-        /**
-         * Get request params for limit & page. Limit is inclusive and page starts at 1.
-         */
-        public limit() {
-            const limitParam = Number.parseInt(this.url.searchParams.get("limit") ?? "100", 10);
-            const pageParam = Number.parseInt(this.url.searchParams.get("page") ?? "1", 10);
-            const limit = Number.isFinite(limitParam) && limitParam >= 0 ? limitParam : 100;
-            const page = Number.isFinite(pageParam) && pageParam > 0 ? pageParam : 1;
-            return {limit, page};
-        }
-    }
-
-    export class Resource {
-        public delete(req: Request): ApiResponse {
-            return new ErrorResponse(405, "The HTTP method `" + req.method + "` is not allowed for endpoint `" + req.url.pathname + "`.");
-        }
-
-        public put(req: Request): ApiResponse {
-            return new ErrorResponse(405, "The HTTP method `" + req.method + "` is not allowed for endpoint `" + req.url.pathname + "`.");
-        }
-
-        public patch(req: Request): ApiResponse {
-            return new ErrorResponse(405, "The HTTP method `" + req.method + "` is not allowed for endpoint `" + req.url.pathname + "`.");
-        }
-    }
-}
-
-export default Api;
