+++ src/response/ApiResponse.ts
@@ -1,7 +1,7 @@
-import Api from "../api/Api.js";
+import ApiRequest from "../api/ApiRequest.js";
 
 export default abstract class ApiResponse {
     protected constructor(public readonly status: number) {
     }
-    public abstract send(req: Api.Request): void | Promise<void>;
+    public abstract send(req: ApiRequest): void | Promise<void>;
 }
